Update network.py
No need to compare all umi1 umi2 in umis. It's redundant. only need to compare unique umis in counts.keys(). Won't be an issue for thousands of umis. but it's a pain in my case, which has hundreds of thousands of umis. Take up a lot memories and time consuming