Fixed get_read_position to be consistent about whether the reported "pos" is included in the read or not.
read.aend is the first position PAST the read, whereas read.pos is the position WITHIN the read. To
be consistent, get_read_position thus needs to substract one from aend to get the START position of
a reversed read (start position in read direction, i.e. the LAST position along the reference that
is part of the read).