Fix exception type in prepare-for-RSEM
prepare-for-rsem tries to identify cases where a read has no mate by catching a look up error. However, the wrong exception type was specified. Since current template is a dictionary of defaultdicts, it shouldn't throw a key error. However the returned list will be empty, so any attempt to index into it will throw an IndexError.